+++ src/test/java/com/maxmind/db/DecoderTest.java
@@ -399,7 +399,7 @@ public class DecoderTest {
 
     @Test
     public void testInvalidControlByte() {
-        ByteBuffer buffer = ByteBuffer.wrap(new byte[] {0x0, 0xF});
+        SingleBuffer buffer = SingleBuffer.wrap(new byte[] {0x0, 0xF});
 
         Decoder decoder = new Decoder(new CHMCache(), buffer, 0);
         InvalidDatabaseException ex = assertThrows(
@@ -418,7 +418,7 @@ public class DecoderTest {
             byte[] input = entry.getValue();
 
             String desc = "decoded " + type.name() + " - " + expect;
-            ByteBuffer buffer = ByteBuffer.wrap(input);
+            SingleBuffer buffer = SingleBuffer.wrap(input);
 
             Decoder decoder = new TestDecoder(cache, buffer, 0);
